refactor: 상태기반 App 스타일 적용 및 리팩토링
