+++ playground/src/App.css
@@ -1,0 +1,193 @@
+/* App.css (바닐라 CSS 버전) */
+
+/* 기본 스타일 */
+body {
+	background-color: #f3f4f6; /* bg-gray-100 */
+	font-family: 'Inter', sans-serif;
+	color: #374151; /* text-gray-800 */
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	min-height: 100vh;
+	padding: 16px; /* p-4 */
+	margin: 0;
+}
+
+#root {
+	width: 100%;
+	max-width: 800px; /* max-w-2xl */
+	margin-left: auto;
+	margin-right: auto;
+}
+
+/* 메인 앱 컨테이너 (부모 레이아웃) */
+.test-app-container {
+	background-color: #ffffff; /* bg-white */
+	box-shadow: 0 20px 25px -5px rgba(0, 0, 0, 0.1), 0 8px 10px -6px
+		rgba(0, 0, 0, 0.1); /* shadow-xl */
+	border-radius: 12px; /* rounded-xl */
+	padding: 32px; /* p-8 */
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	gap: 24px; /* space-y-6 */
+	border: 1px solid #e5e7eb; /* border border-gray-200 */
+}
+
+/* 카운트 표시 */
+.count-display {
+	font-size: 60px; /* text-5xl (약 60px) */
+	font-weight: 800; /* font-extrabold */
+	color: #2563eb; /* text-blue-600 */
+	margin-bottom: 16px; /* mb-4 */
+}
+
+/* 앱 버튼 */
+.app-button {
+	padding: 12px 24px; /* px-6 py-3 */
+	border-radius: 9999px; /* rounded-full */
+	font-size: 20px; /* text-lg */
+	font-weight: 600; /* font-semibold */
+	transition: all 0.3s ease-in-out; /* transition duration-300 ease-in-out */
+	border: none;
+	cursor: pointer;
+	outline: none; /* focus:outline-none */
+}
+
+.app-button:hover {
+	transform: scale(1.05); /* hover:scale-105 */
+}
+
+.app-button:focus {
+	box-shadow: 0 0 0 4px rgba(59, 130, 246, 0.3); /* focus:ring-4 (tailwind default blue-300) */
+}
+
+.app-button.primary {
+	background-color: #3b82f6; /* bg-blue-500 */
+	color: #ffffff; /* text-white */
+}
+
+.app-button.primary:hover {
+	background-color: #2563eb; /* hover:bg-blue-600 */
+}
+
+.app-button.primary:focus {
+	box-shadow: 0 0 0 4px rgba(99, 102, 241, 0.3); /* focus:ring-blue-300 */
+}
+
+.app-button.secondary {
+	background-color: #9333ea; /* bg-purple-500 */
+	color: #ffffff; /* text-white */
+}
+
+.app-button.secondary:hover {
+	background-color: #7c2d96; /* hover:bg-purple-600 */
+}
+
+.app-button.secondary:focus {
+	box-shadow: 0 0 0 4px rgba(167, 139, 250, 0.3); /* focus:ring-purple-300 */
+}
+
+/* 내부 카운터 컨테이너 (자식 레이아웃) */
+.inner-counter-container {
+	background-color: #f9fafb; /* bg-gray-50 */
+	border: 1px solid #e5e7eb; /* border border-gray-200 */
+	border-radius: 8px; /* rounded-lg */
+	padding: 24px; /* p-6 */
+	margin-top: 32px; /* mt-8 */
+	display: flex;
+	flex-direction: column;
+	align-items: center; /* items-center */
+	gap: 16px; /* space-y-4 */
+	text-align: center;
+	width: 100%;
+	max-width: 384px; /* max-w-sm */
+}
+
+.inner-count-display {
+	font-size: 36px; /* text-4xl */
+	font-weight: 700; /* font-bold */
+	color: #16a34a; /* text-green-600 */
+}
+
+.inner-button {
+	background-color: #22c55e; /* bg-green-500 */
+	color: #ffffff; /* text-white */
+	padding: 8px 20px; /* px-5 py-2 */
+	border-radius: 9999px; /* rounded-full */
+	font-weight: 500; /* font-medium */
+	transition: all 0.3s ease-in-out; /* transition duration-300 ease-in-out */
+	border: none;
+	cursor: pointer;
+	outline: none; /* focus:outline-none */
+}
+
+.inner-button:hover {
+	background-color: #16a34a; /* hover:bg-green-600 */
+}
+
+.inner-button:focus {
+	box-shadow: 0 0 0 4px rgba(52, 211, 153, 0.3); /* focus:ring-green-300 */
+}
+
+.inner-counter-container span:last-child {
+	font-size: 14px; /* text-sm */
+	color: #6b7280; /* text-gray-500 */
+	margin-top: 8px; /* mt-2 */
+}
+
+/* 모달 오버레이 (조건부 렌더링 - 모달 배경) */
+.modal-overlay {
+	position: fixed;
+	top: 0;
+	left: 0;
+	right: 0;
+	bottom: 0;
+	background-color: rgba(0, 0, 0, 0.5); /* bg-black bg-opacity-50 */
+	z-index: 40;
+}
+
+/* 모달 컨테이너 (조건부 렌더링 - 모달 본체) */
+.modal-container {
+	position: fixed;
+	top: 50%;
+	left: 50%;
+	transform: translate(-50%, -50%); /* -translate-x-1/2 -translate-y-1/2 */
+	background-color: #ffffff; /* bg-white */
+	box-shadow: 0 25px 50px -12px rgba(0, 0, 0, 0.25); /* shadow-2xl */
+	border-radius: 8px; /* rounded-lg */
+	padding: 32px; /* p-8 */
+	z-index: 50;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	gap: 24px; /* space-y-6 */
+}
+
+/* 모달 버튼 */
+.modal-button {
+	background-color: #ef4444; /* bg-red-500 */
+	color: #ffffff; /* text-white */
+	padding: 12px 24px; /* px-6 py-3 */
+	border-radius: 9999px; /* rounded-full */
+	font-weight: 600; /* font-semibold */
+	transition: all 0.3s ease-in-out; /* transition duration-300 ease-in-out */
+	border: none;
+	cursor: pointer;
+	outline: none; /* focus:outline-none */
+}
+
+.modal-button:hover {
+	background-color: #dc2626; /* hover:bg-red-600 */
+}
+
+.modal-button:focus {
+	box-shadow: 0 0 0 4px rgba(248, 113, 113, 0.3); /* focus:ring-red-300 */
+}
+
+/* 모달 내용 */
+.modal-content {
+	font-size: 24px; /* text-2xl */
+	font-weight: 600; /* font-semibold */
+	color: #4b5563; /* text-gray-700 */
+}
